AI-Disease/requirements.txt
fastapi
uvicorn
python-multipart
ultralytics
pillow
pillow-heif
torch 
torchvision
opencv-python-headless